+++ src/Symfony/Bundle/FrameworkBundle/Tests/Translation/TranslatorTest.php
@@ -94,13 +94,14 @@ class TranslatorTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals('foobarbax (sr@latin)', $translator->trans('foobarbax'));
     }
 
+    /**
+     * @expectedException \InvalidArgumentException
+     */
     public function testTransWithCachingWithInvalidLocale()
     {
         $loader = $this->getMock('Symfony\Component\Translation\Loader\LoaderInterface');
         $translator = $this->getTranslator($loader, array('cache_dir' => $this->tmpDir), 'loader', '\Symfony\Bundle\FrameworkBundle\Tests\Translation\TranslatorWithInvalidLocale');
-        $translator->setLocale('invalid locale');
 
-        $this->setExpectedException('\InvalidArgumentException');
         $translator->trans('foo');
     }
 
@@ -302,8 +303,8 @@ class TranslatorWithInvalidLocale extends Translator
     /**
      * {@inheritdoc}
      */
-    public function setLocale($locale)
+    public function getLocale()
     {
-        $this->locale = $locale;
+        return 'invalid locale';
     }
 }